consolidate swap
