+++ packages/web/src/sections/swap/lifi-status-modal.tsx
@@ -1,9 +1,6 @@
 'use client';
 
 import { useTranslate } from '@/locales';
-import { useState, useEffect } from 'react';
-import { buildAuthHeaders } from '@/utils/http';
-import { ETH_RPC_URL, SOL_RPC_URL } from '@/hooks/use-chain-portfolio';
 
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
@@ -14,202 +11,40 @@ import CircularProgress from '@mui/material/CircularProgress';
 
 import { Iconify } from '@/components/template/iconify';
 
-// ---------------------------------------------------------------------------
-// Staged tracker for a cross-chain (LI.FI) swap. Opens the instant the source
-// tx is broadcast and walks through: Submitted → Confirming on source chain →
-// Bridging → Completed/Failed. Non-blocking (closable), records the swap only
-// after the source tx confirms (so a dropped tx never becomes a phantom row),
-// and logs every step to the console.
-// ---------------------------------------------------------------------------
+import { chainName } from './engines/lifi-tracker';
 
-const CHAIN = {
-  ETH: 1,
-  SOL: 1151111081099710,
-  BTC: 20000000000001,
-};
+import type { Stage } from './engines/lifi-tracker';
 
-function chainName(id: number): string {
-  if (id === CHAIN.ETH) return 'Ethereum';
-  if (id === CHAIN.SOL) return 'Solana';
-  if (id === CHAIN.BTC) return 'Bitcoin';
-  return 'source chain';
-}
+// ---------------------------------------------------------------------------
+// Presentational view of a cross-chain (LI.FI) swap's progress: Submitted →
+// Confirming on source chain → Bridging → Completed/Failed/Refunded. The actual
+// tracking (confirm + record + bridge poll + balance refresh) runs in the swap
+// engine via useLifiTracker, so closing this dialog never stops the swap — it
+// just hides the view. `stage` is fed in from that tracker.
+// ---------------------------------------------------------------------------
 
 interface LifiStatusModalProps {
   open: boolean;
   onClose: () => void;
+  stage: Stage;
   txHash: string;
   fromChainId: number;
   toChainId: number;
   fromSymbol: string;
   toSymbol: string;
-  amountIn: string;
-  amountOut: string;
 }
-
-type Stage = 'confirming' | 'bridging' | 'done' | 'failed' | 'refunded';
-
-const log = (msg: string, extra?: unknown) =>
-  extra !== undefined ? console.log(`[lifi-swap] ${msg}`, extra) : console.log(`[lifi-swap] ${msg}`);
-
-function delay(ms: number) {
-  return new Promise((r) => setTimeout(r, ms));
-}
-
-// --- source-chain confirmation -------------------------------------------------
-
-async function rpc(url: string, method: string, params: unknown[]): Promise<any> {
-  const res = await fetch(url, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ jsonrpc: '2.0', id: 1, method, params }),
-  });
-  if (!res.ok) throw new Error(`rpc ${method} ${res.status}`);
-  const data = await res.json();
-  if (data.error) throw new Error(data.error.message ?? method);
-  return data.result;
-}
-
-/** 'confirmed' | 'reverted' | 'dropped' */
-async function confirmSource(
-  fromChainId: number,
-  txHash: string,
-  stop: () => boolean
-): Promise<'confirmed' | 'reverted' | 'dropped'> {
-  // Bitcoin: broadcast acceptance is enough — the bridge watches the mempool.
-  if (fromChainId === CHAIN.BTC) return 'confirmed';
-
-  if (fromChainId === CHAIN.ETH) {
-    for (let i = 0; i < 45 && !stop(); i++) {
-      try {
-        const receipt = await rpc(ETH_RPC_URL, 'eth_getTransactionReceipt', [txHash]);
-        if (receipt) {
-          log('ETH receipt status', receipt.status);
-          return receipt.status === '0x1' ? 'confirmed' : 'reverted';
-        }
-      } catch (e) {
-        log('ETH receipt poll error', (e as Error).message);
-      }
-      await delay(4000);
-    }
-    return 'dropped';
-  }
-
-  if (fromChainId === CHAIN.SOL) {
-    for (let i = 0; i < 30 && !stop(); i++) {
-      try {
-        const res = await rpc(SOL_RPC_URL, 'getSignatureStatuses', [[txHash]]);
-        const st = res?.value?.[0];
-        if (st) {
-          if (st.err) return 'reverted';
-          if (st.confirmationStatus === 'confirmed' || st.confirmationStatus === 'finalized') {
-            return 'confirmed';
-          }
-        }
-      } catch (e) {
-        log('SOL status poll error', (e as Error).message);
-      }
-      await delay(2000);
-    }
-    return 'dropped';
-  }
-
-  return 'confirmed';
-}
-
-async function pollBridge(
-  txHash: string,
-  fromChainId: number,
-  toChainId: number,
-  stop: () => boolean
-): Promise<'DONE' | 'FAILED' | 'REFUNDED' | null> {
-  for (let i = 0; i < 80 && !stop(); i++) {
-    try {
-      const res = await fetch(
-        `/api/lifi/status?txHash=${txHash}&fromChain=${fromChainId}&toChain=${toChainId}`
-      );
-      if (res.ok) {
-        const data = await res.json();
-        const s = data?.status?.status as string | undefined;
-        const sub = String(data?.status?.substatus ?? '').toUpperCase();
-        log('bridge status', `${s} / ${sub}`);
-        // A refund is reported as DONE+REFUNDED (or PARTIAL) — funds returned,
-        // destination not delivered.
-        if (s === 'DONE') return sub === 'REFUNDED' || sub === 'PARTIAL' ? 'REFUNDED' : 'DONE';
-        if (s === 'FAILED' || s === 'INVALID') return 'FAILED';
-      }
-    } catch (e) {
-      log('bridge poll error', (e as Error).message);
-    }
-    await delay(15000);
-  }
-  return null;
-}
-
-// ---------------------------------------------------------------------------
 
 export function LifiStatusModal({
   open,
   onClose,
+  stage,
   txHash,
   fromChainId,
   toChainId,
   fromSymbol,
   toSymbol,
-  amountIn,
-  amountOut,
 }: LifiStatusModalProps) {
   const { t } = useTranslate();
-  const [stage, setStage] = useState<Stage>('confirming');
-
-  useEffect(() => {
-    if (!open || !txHash) return undefined;
-    let cancelled = false;
-    const stop = () => cancelled;
-    setStage('confirming');
-
-    (async () => {
-      log(`Submitted ${fromSymbol}→${toSymbol}`, txHash);
-
-      // 1) Confirm the source-chain transaction actually landed.
-      const src = await confirmSource(fromChainId, txHash, stop);
-      if (cancelled) return;
-      if (src !== 'confirmed') {
-        log('source not confirmed', src);
-        setStage('failed');
-        return;
-      }
-      log('source confirmed');
-
-      // 2) Record now (source landed) so it shows as one Swap row, then refresh feeds.
-      try {
-        const headers = await buildAuthHeaders();
-        await fetch('/api/lifi/record', {
-          method: 'POST',
-          headers: { ...headers, 'Content-Type': 'application/json' },
-          body: JSON.stringify({ fromSymbol, toSymbol, amountIn, amountOut, txHash }),
-        });
-        log('recorded swap');
-      } catch (e) {
-        log('record failed (non-fatal)', (e as Error).message);
-      }
-      window.dispatchEvent(new Event('nf:activity-updated'));
-
-      // 3) Track the bridge until it delivers.
-      setStage('bridging');
-      const bridge = await pollBridge(txHash, fromChainId, toChainId, stop);
-      if (cancelled) return;
-      if (bridge === 'DONE') setStage('done');
-      else if (bridge === 'REFUNDED') setStage('refunded');
-      else if (bridge === 'FAILED') setStage('failed');
-      window.dispatchEvent(new Event('nf:activity-updated'));
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [open, txHash]);
 
   const steps: { key: Stage | 'submitted'; label: string }[] = [
     { key: 'submitted', label: t('Submitted') },
